+++ R/OptimizerRandomSearch.R
@@ -20,6 +20,8 @@
 #' Maximum number of points to try in a batch.}
 #' }
 #'
+#' @template section_progress_bars
+#' 
 #' @source
 #' `r format_bib("bergstra_2012")`
 #'