+++ NEWS.md
@@ -1,4 +1,4 @@
-# bbotk 0.2.2.9000
+# bbotk 0.3.0
 
 * Improved `TerminatorPerfRearched` documentation.
 * Added `check_values` flag in `OptimInstance`, `OptimInstanceSingleCrit`,